+++ dotnet/TF2SA.Http/LogsTF/Config/LogsTFConfig.cs
@@ -4,5 +4,5 @@ public class LogsTFConfig
 {
 	public const string LogsTFConfigSection = "LogsTF";
 	public string BaseUrl { get; set; } = string.Empty;
-	public uint[] Uploaders { get; set; } = Array.Empty<uint>();
+	public ulong[] Uploaders { get; set; } = Array.Empty<ulong>();
 }
